Migrate TOSCA parser from xopera-opera repository
This commit brings big changes as we're taking TOSCA YAML parser from
opera TOSCA orchestrator and migrating it here (as stated in
https://github.com/xlab-si/xopera-opera/issues/170).

The parser's code is not just migrated, but also improved with typing
and docstrings (in reStructuredText format, so we could document every
piece of it if needed). There were also some minor modifications of the
Python functions because there are some files that contain parts that
belong to parser and some sections that should go to the orchestrator
(these sections will be added in the orchestrator by extending existing
parser Python classes).